Fix unexpected brushReset behavior.
Calling brush() on brushReset leads to __.brushed being set to an empty
array. As a result, a call to render() would lead to an empty ||-coords.
This commit implements (AFAICS) the correct behavior for brushReset. It
sets __.brushed to false and calls render() in stead of brush() when the
g.dimensions are updated.