Rename the pinch canvas layer to strum and style it in the code.
I had hoped of getting rid of the layer altogether, however I need to
"install" the drag behavior on it. I tried to install it on the svg element
without luck and installing it on g#strums doesn't work either because it's
size is based on its content, so it won't "see" any drag events.

The styling of the strums canvas, which enables listening to mouse-vents,
must be done in the code (as opposed to the css) for reasons mentioned in
pinchable.js.