Implement OR brushing for strums.
This required, among other things, to make strums 'global'. That is, it is
now visible in the full module, in stead of being passed as a parameter. A
downside of this is, that I don't know how generic this is when you would
like to reuse a d3.parcoords object for two or more parallel coordinates
instances. However, this doesn't seem to be supported at this point anyhow.