Make sure that upper and lower boundaries are respected as well.
THe left and right boundaries (i.e. left and right axes of an segment)
were respected. That is, a strum can only be placed between two axes and
not start in one segment and end in the next. This wasn't true yet for
upper and lower boundaries. This is fixed now, and a strum can not end
above or below the axes any longer.