Make sure that shadows are re-rendered on axes reordering.
I'm not a big fan of the shadows functionality, especially because it's
current implementation is not suitable for a large number of items.
However, if it's used, it should be at least working correctly. Before this
fix, the shadows wouldn't update on axis reordering, which gave a strange
(and incorrect) effect.