Refine axis flipping even more.
In stead of updating all axes on an axis flip, we now just update the one
that is affected. This avoids confusing behavior which I observed when
first swapping order of axes and next flipping one.

I also removed code duplication by moving out this function into a seperate
one which is reused in createAxes and updateAxes.