Changed reorder()'s sort to be stable
The Array.sort() function in reorder() can be unstable depending on the
browser. This means that if some of the values passed in to the
reorder() function result in having the same pixel height, their
respective dimensions might be ordered differently everytime reorder()
is called. To solve this issue, the dimensions are ordered on name
whenever such a conflict occurs.