Added the brushedColor and alphaOnBrushed options
brushedColor() works the same way as color() but it will only color the
brushed items. When brushedColor is set to null (default) it will use
color instead.
alphaOnBrushed() sets the alpha of the data items when brushing takes
place. This can be used to render both data and brushed at the same time
but putting extra emphasize on the brushed items.

To achieve this a new canvas layer was added for the brushed items.
Thanks to this new layer it is now possible to rerender the brushed
items independently from the rest of the data. This can be done with the
renderBrushed function.